+++ CHANGES.rst
@@ -1,6 +1,18 @@
 Changelog
 =========
 
+1.0.2 (2019-05-20)
+------------------
+
+- New behavior ``IPreviewImage`` (requires ``plone.namedfile``)
+
+  - with (rudimentary, for now) ``getThumbnailPath`` method
+
+- Translations
+
+[tobiasherp]
+
+
 1.0.1 (2018-10-17)
 ------------------